Move BlackRendererSettings to `render.py` as asked
